requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-speech==1.36.0
azure-core==1.30.1
azure-storage-blob==12.19.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dataclasses-json==0.6.7
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.10
django-allauth==0.58.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
duplicity==0.8.21
environs==9.5.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fasteners==0.14.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.6.1
future==0.18.2
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.63.0
gunicorn==21.2.0
gyp==0.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.6
importlib-metadata==4.6.4
inflection==0.5.1
isodate==0.6.1
jeepney==0.7.1
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-cloud==0.0.9
llama-index==0.10.55
llama-index-agent-openai==0.2.8
llama-index-cli==0.1.12
llama-index-core==0.10.55
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.6
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-index-readers-web==0.1.23
llama-parse==0.4.9
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.21.3
mdurl==0.1.2
milvus-lite==2.4.8
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
openai==1.36.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
paramiko==2.9.3
pbkdf2==1.3
pexpect==4.8.0
pillow==10.4.0
pjsua2==2.14.dev0
platformdirs==4.2.0
playwright==1.45.0
protobuf==4.25.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
pymilvus==2.4.4
PyNaCl==1.5.0
pyparsing==3.1.1
pypdf==4.3.0
PyPDF2==3.0.1
pyRFC3339==1.1
PySocks==1.7.1
pyst==0.9
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
railway==0.0.4
regex==2024.5.15
reportlab==3.6.8
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
selenium==4.23.0
sentence-transformers==3.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spider-client==0.0.27
SQLAlchemy==2.0.31
sqlparse==0.4.4
striprtf==0.0.26
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
together==1.2.2
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
trio==0.26.0
trio-websocket==0.11.1
triton==2.3.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.1.0
usb-creator==0.3.7
virtualenv==20.25.1
wadllib==1.3.6
websocket-client==1.8.0
websockets==12.0
wifi==0.3.8
wrapt==1.16.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
